#6: Fix File Venue Item Table View Cell
