+++ FourSquare/Controllers/Home/HomeViewController.swift
@@ -45,7 +45,7 @@ class HomeViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         self.setDefaultMenuItems()
-        self.setUpMenuPage()
+        self.setUpMenuPage(isDefault: true)
         self.setUpNotificationCenter()
     }
 
@@ -73,8 +73,11 @@ class HomeViewController: UIViewController {
         }
     }
 
-    private func setUpMenuPage() {
-        let menuItemWidth = (kScreenSize.width - 20) / 3
+    private func setUpMenuPage(isDefault isDefault: Bool) {
+        var menuItemWidth = (kScreenSize.width - 20) / 3
+        if !isDefault {
+            menuItemWidth = (kScreenSize.width - 20) / 3.5
+        }
         let parameters: [CAPSPageMenuOption] = [.MenuHeight(35),
                 .MenuMargin(5),
                 .MenuItemWidth(menuItemWidth),
@@ -107,7 +110,11 @@ class HomeViewController: UIViewController {
                 pageViewController.menuItem = activeMenuItem
                 self.itemViewControllers.append(pageViewController)
             }
-            self.setUpMenuPage()
+            if newActiveMenuItems.count == 0 {
+                self.setUpMenuPage(isDefault: true)
+                return
+            }
+            self.setUpMenuPage(isDefault: false)
         }
     }
