+++ FourSquare/Controllers/Page/VenueItemTableViewCell/VenueItemTableViewCell.swift
@@ -1,0 +1,24 @@
+//
+//  VenueItemTableViewCell.swift
+//  FourSquare
+//
+//  Created by Mylo Ho on 8/4/16.
+//  Copyright © 2016 Le Van Binh. All rights reserved.
+//
+
+import UIKit
+
+class VenueItemTableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        self.selectionStyle = .None
+    }
+
+    override func setSelected(selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
